Add FreeATS
 ### FreeATS

FreeATS is an open source applicant tracking system built on Ruby on Rails,
offering self-hosting capabilities and an active community.
It provides an easy start with a free trial of FreeATS Cloud and allows users to host
the system on their own servers, ensuring full control over data and infrastructure.
The project is particularly relevant in today's job market, offering flexibility and control
for modern HR professionals and companies looking to optimize their hiring processes.

- [FreeATS Repository](https://github.com/freeats/freeats/)
- [FreeATS Website](https://www.freeats.com/)